requirements.txt
# TrueCheck — Requirements
# LLM dependencies (openai, google-generativeai, google-genai) fully removed.

# Web framework
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0

# Firebase
firebase-admin==6.7.0
google-cloud-firestore==2.20.1

# ML pipeline
scikit-learn>=1.4.0
numpy>=1.26.0
joblib>=1.3.0

# Data fetching / HTTP
requests==2.32.3

# Environment
python-dotenv>=1.0.0
